Fixed issue with files being deleted when changing the casing of a file name.
When using integrate/add/delete the source and destination can't be the
same.
Background: This happens when a files casing is changed. However, the
asset list doesn't reflect that case change. As such it's an issue with
the Unity editor i.e. it shouldn't issue NOP commands. Also, a case
change should only be reflected in Perforce if the server is set to
being case sensitive.